# apps/hs-blog-server/admin-openapi.json
{"openapi":"3.0.0","paths":{"/api/admin/openai/chat":{"post":{"operationId":"chat","parameters":[],"responses":{"201":{"description":""}},"tags":["admin","大模型应用"]}},"/api/admin/openai/analyze-code":{"post":{"operationId":"analyzeCode","parameters":[],"responses":{"201":{"description":""}},"summary":"代码分析","tags":["admin","大模型应用"]}},"/api/admin/openai/generate-short-content":{"post":{"description":"生成不超过300个汉字的简短摘要，准确概括文章核心内容","operationId":"generateShortContent","parameters":[],"responses":{"201":{"description":""}},"summary":"生成短内容","tags":["admin","大模型应用"]}},"/api/admin/openai/generate-long-content":{"post":{"description":"生成博客长对话内容","operationId":"generateLongContent","parameters":[],"responses":{"201":{"description":""}},"summary":"生成长对话内容","tags":["admin","大模型应用"]}},"/api/admin/openai/generate-short-audio":{"post":{"operationId":"generateShortAudio","parameters":[],"responses":{"201":{"description":""}},"summary":"tts-生成短文本音频","tags":["admin","大模型应用"]}},"/api/admin/openai/generate-long-audio":{"post":{"operationId":"generateLongAudio","parameters":[],"responses":{"201":{"description":""}},"summary":"tts-生成长对话音频","tags":["admin","大模型应用"]}},"/api/admin/oss/ali/article-img":{"post":{"operationId":"uploadArticleFile","parameters":[],"responses":{"201":{"description":""}},"tags":["admin","阿里云OSS"]},"delete":{"operationId":"deleteDirectory","parameters":[],"responses":{"200":{"description":""}},"tags":["admin","阿里云OSS"]}},"/api/admin/oss/ali/update-article-id":{"post":{"operationId":"updateArticleId","parameters":[],"responses":{"201":{"description":""}},"tags":["admin","阿里云OSS"]}},"/api/admin/oss/ali/upload-audio":{"post":{"operationId":"uploadAudio","parameters":[],"responses":{"201":{"description":""}},"tags":["admin","阿里云OSS"]}},"/api/admin/article/list":{"get":{"operationId":"findAll","parameters":[],"responses":{"200":{"description":""}},"security":[{"bearer":[]}],"summary":"获取文章列表","tags":["admin","文章模块"]}},"/api/admin/article/status":{"put":{"operationId":"editArticlesStatus","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/EditArticlesStatus"}}}},"responses":{"200":{"description":""}},"security":[{"bearer":[]}],"summary":"更新文章状态","tags":["admin","文章模块"]}},"/api/admin/article":{"put":{"operationId":"updateArticle","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateArticleDto"}}}},"responses":{"200":{"description":""}},"security":[{"bearer":[]}],"summary":"更新文章","tags":["admin","文章模块"]},"delete":{"operationId":"delete","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/DeleteArticlesDto"}}}},"responses":{"200":{"description":""}},"security":[{"bearer":[]}],"summary":"删除文章","tags":["admin","文章模块"]},"post":{"operationId":"createArticle","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateArticleDto"}}}},"responses":{"201":{"description":""}},"security":[{"bearer":[]}],"tags":["admin","文章模块"]}},"/api/admin/article/details":{"get":{"operationId":"articleDetails","parameters":[{"name":"id","required":true,"in":"query","schema":{"type":"number"}}],"responses":{"200":{"description":""}},"security":[{"bearer":[]}],"summary":"获取文章详情","tags":["admin","文章模块"]}},"/api/admin/article/tags":{"get":{"operationId":"findAllTags","parameters":[],"responses":{"200":{"description":""}},"security":[{"bearer":[]}],"tags":["admin","文章模块"]}},"/api/admin/article/update-publish-time":{"post":{"operationId":"updatePublishTime","parameters":[],"responses":{"200":{"description":""}},"security":[{"bearer":[]}],"summary":"更新文章发布时间","tags":["admin","文章模块"]}},"/api/admin/article/tts":{"put":{"operationId":"updateTts","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateArticleTtsDto"}}}},"responses":{"200":{"description":""}},"security":[{"bearer":[]}],"summary":"更新文章TTS","tags":["admin","文章模块"]}},"/api/admin/article/stats/trend":{"get":{"operationId":"getArticlePublishTrend","parameters":[],"responses":{"200":{"description":""}},"security":[{"bearer":[]}],"summary":"获取文章发布趋势","tags":["admin","文章模块"]}},"/api/admin/article/stats/category":{"get":{"operationId":"getArticleCategoryStats","parameters":[],"responses":{"200":{"description":""}},"security":[{"bearer":[]}],"summary":"获取文章分类统计","tags":["admin","文章模块"]}},"/api/admin/article/stats/hot":{"get":{"operationId":"getHotArticles","parameters":[],"responses":{"200":{"description":""}},"security":[{"bearer":[]}],"summary":"获取热门文章统计","tags":["admin","文章模块"]}},"/api/admin/article/stats/time-distribution":{"get":{"operationId":"getArticleTimeDistribution","parameters":[],"responses":{"200":{"description":""}},"security":[{"bearer":[]}],"summary":"获取文章发布时间分布","tags":["admin","文章模块"]}},"/api/admin/article/stats/overview":{"get":{"operationId":"getArticleOverview","parameters":[],"responses":{"200":{"description":""}},"security":[{"bearer":[]}],"summary":"获取文章数量概览","tags":["admin","文章模块"]}},"/api/admin/article/stats/words":{"get":{"operationId":"getArticleWordsDistribution","parameters":[],"responses":{"200":{"description":""}},"security":[{"bearer":[]}],"summary":"获取文章字数统计","tags":["admin","文章模块"]}},"/api/admin/category":{"post":{"operationId":"createCategory","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateCategoryDto"}}}},"responses":{"201":{"description":""}},"security":[{"bearer":[]}],"summary":"新增分类","tags":["admin","分类模块"]},"get":{"operationId":"findAll","parameters":[],"responses":{"200":{"description":""}},"security":[{"bearer":[]}],"summary":"查询所有分类","tags":["admin","分类模块"]},"put":{"operationId":"updateCategory","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateCategoryDto"}}}},"responses":{"200":{"description":""}},"security":[{"bearer":[]}],"summary":"更新分类","tags":["admin","分类模块"]},"delete":{"operationId":"delete","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/DeleteCategoryDto"}}}},"responses":{"200":{"description":""}},"security":[{"bearer":[]}],"summary":"删除分类","tags":["admin","分类模块"]}},"/api/admin/category/{id}":{"get":{"operationId":"findOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":""}},"security":[{"bearer":[]}],"summary":"查询单个分类","tags":["admin","分类模块"]}},"/api/admin/tag/stats":{"get":{"operationId":"getTagsStats","parameters":[],"responses":{"200":{"description":""}},"security":[{"bearer":[]}],"summary":"标签统计","tags":["admin","标签管理"]}},"/api/admin/tag/trend":{"get":{"operationId":"getTagsTrend","parameters":[],"responses":{"200":{"description":""}},"security":[{"bearer":[]}],"summary":"标签趋势","tags":["admin","标签管理"]}},"/api/admin/tag/relation":{"get":{"operationId":"getTagsRelation","parameters":[],"responses":{"200":{"description":""}},"security":[{"bearer":[]}],"summary":"标签关系","tags":["admin","标签管理"]}}},"info":{"title":"hs blog API","description":"hs blog 项目 API 文档","version":"1.0.0","contact":{"name":"火山","url":"https://github.com/huoshan25/sky-hub","email":"2633057734@qq.com"}},"tags":[],"servers":[],"components":{"securitySchemes":{"bearer":{"scheme":"bearer","bearerFormat":"JWT","type":"http","name":"Authorization","description":"请输入 JWT 格式的 Token","in":"header"}},"schemas":{"SendEmailCodeDto":{"type":"object","properties":{"email":{"type":"string","description":"邮箱地址","example":"example@sky.com"}},"required":["email"]},"VerifyEmailCodeDto":{"type":"object","properties":{"email":{"type":"string","description":"邮箱地址","example":"example@sky.com"},"code":{"type":"string","description":"验证码","example":"123456","minLength":6,"maxLength":6}},"required":["email","code"]},"ApiResponseBase":{"type":"object","properties":{}},"registerResVo":{"type":"object","properties":{"accessToken":{"type":"string","description":"用户token"},"refreshToken":{"type":"string","description":"刷新token"}},"required":["accessToken","refreshToken"]},"RegisterDto":{"type":"object","properties":{"username":{"type":"string","description":"用户名","example":"skyuser","minLength":3},"email":{"type":"string","description":"邮箱地址","example":"example@sky.com"},"password":{"type":"string","description":"密码","example":"password123","minLength":6},"confirmPassword":{"type":"string","description":"确认密码","example":"password123","minLength":6},"code":{"type":"string","description":"邮箱验证码","example":"123456"}},"required":["username","email","password","confirmPassword","code"]},"LoginResponseVo":{"type":"object","properties":{"accessToken":{"type":"string","description":"用户token"},"refreshToken":{"type":"string","description":"刷新token"}},"required":["accessToken","refreshToken"]},"LoginDto":{"type":"object","properties":{"usernameOrEmail":{"type":"string","description":"用户名或邮箱","example":"skyuser"},"password":{"type":"string","description":"密码","example":"password123"}},"required":["usernameOrEmail","password"]},"LoginResponseDto":{"type":"object","properties":{"accessToken":{"type":"string","description":"访问令牌"},"refreshToken":{"type":"string","description":"刷新令牌"},"expiresIn":{"type":"number","description":"访问令牌过期时间（秒）"}},"required":["accessToken","refreshToken","expiresIn"]},"RefreshTokenDto":{"type":"object","properties":{"refreshToken":{"type":"string","description":"刷新令牌","example":"eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."}},"required":["refreshToken"]},"userVo":{"type":"object","properties":{"id":{"type":"number","description":"用户id"},"userName":{"type":"string","description":"用户名"},"email":{"type":"string","description":"邮箱"},"avatar":{"type":"string","description":"头像"},"bio":{"type":"string","description":"个人简介"}},"required":["id","userName","email","avatar","bio"]},"UpdateUserDto":{"type":"object","properties":{"userName":{"type":"string","description":"用户名","example":"johndoe"},"avatar":{"type":"string","description":"头像URL","example":"https://example.com/avatar.jpg"},"bio":{"type":"string","description":"个人简介","example":"这是一段个人简介"}}},"EditArticlesStatus":{"type":"object","properties":{}},"UpdateArticleDto":{"type":"object","properties":{}},"DeleteArticlesDto":{"type":"object","properties":{}},"CreateArticleDto":{"type":"object","properties":{}},"CreateArticleTtsDto":{"type":"object","properties":{}},"CreateCategoryDto":{"type":"object","properties":{}},"UpdateCategoryDto":{"type":"object","properties":{}},"DeleteCategoryDto":{"type":"object","properties":{}}}}}